[CompileOpt] Refine enforce code and remove boost/variant include (#41093)
* refine enforce code

* refine enforce code

* fix compile failed

* fix infrt failed
