+++ paddle/fluid/framework/type_defs.h
@@ -21,8 +21,8 @@ limitations under the License. */
 #include <unordered_map>
 #include <unordered_set>
 #include <vector>
+#include "paddle/fluid/imperative/type_defs.h"
 #include "paddle/fluid/platform/variant.h"
-#include "paddle/phi/core/compat/type_defs.h"
 #include "paddle/utils/small_vector.h"
 
 namespace paddle {
@@ -38,6 +38,12 @@ class InferNoNeedBufferVarsFN;
 // TODO(panyx0718): Replace vector with something like gtl::Vector.
 using VariableNameMap = std::map<std::string, std::vector<std::string>>;
 using VariableValueMap = std::map<std::string, std::vector<Variable*>>;
+
+using Attribute = boost::variant<
+    boost::blank, int, float, std::string, std::vector<int>, std::vector<float>,
+    std::vector<std::string>, bool, std::vector<bool>, BlockDesc*, int64_t,
+    std::vector<BlockDesc*>, std::vector<int64_t>, std::vector<double>>;
+using AttributeMap = std::unordered_map<std::string, Attribute>;
 
 #ifdef PADDLE_WITH_ASCEND_CL
 using NPUAttribute =
