+++ paddle/fluid/memory/allocation/allocator.h
@@ -33,8 +33,8 @@ namespace allocation {
 // Exception when `Alloc`/`AllocShared` failed
 struct BadAlloc : public std::exception {
   inline explicit BadAlloc(std::string err_msg, const char* file, int line)
-      : err_str_(platform::GetTraceBackString(std::move(err_msg), file, line)) {
-  }
+      : err_str_(platform::GetCompleteTraceBackString(std::move(err_msg), file,
+                                                      line)) {}
 
   const char* what() const noexcept override { return err_str_.c_str(); }
