+++ paddle/phi/core/CMakeLists.txt
@@ -6,7 +6,7 @@ set(phi_enforce_deps errors flags)
 if(WITH_GPU)
   set(phi_enforce_deps ${phi_enforce_deps} external_error_proto)
 endif()
-cc_library(phi_enforce INTERFACE SRCS enforce.cc DEPS ${phi_enforce_deps})
+cc_library(phi_enforce SRCS enforce.cc DEPS ${phi_enforce_deps})
 
 cc_library(kernel_factory SRCS kernel_factory.cc DEPS phi_enforce fluid_convert_utils)
 cc_library(kernel_context SRCS kernel_context.cc DEPS phi_enforce phi_context)
