+++ paddle/phi/core/enforce.h
@@ -28,38 +28,32 @@ limitations under the License. */
 #include <windows.h>  // GetModuleFileName, Sleep
 #endif
 
-#include <fstream>
-#include <iomanip>
-#include <memory>
 #include <sstream>
 #include <stdexcept>
 #include <string>
 #include <type_traits>
-#include <utility>
 
 #if !defined(_WIN32) && !defined(PADDLE_WITH_MUSL)
 #include <execinfo.h>
 #endif
 
-#define GLOG_NO_ABBREVIATED_SEVERITIES  // msvc conflict logging with windows.h
 #include "gflags/gflags.h"
-#include "glog/logging.h"
+
+#define GLOG_NO_ABBREVIATED_SEVERITIES  // msvc conflict logging with windows.h
 #include "paddle/phi/core/errors.h"
 #include "paddle/utils/string/printf.h"
 #include "paddle/utils/string/to_string.h"
 
-// Note: these headers for simplify demangle type string
-#include "paddle/phi/core/compat/type_defs.h"
+DECLARE_int32(call_stack_level);
 
 namespace phi {
 class ErrorSummary;
 }  // namespace phi
 
-DECLARE_int32(call_stack_level);
 namespace phi {
 namespace enforce {
-/** HELPER MACROS AND FUNCTIONS **/
 
+/** HELPER MACROS AND FUNCTIONS **/
 #ifndef PADDLE_MAY_THROW
 #define PADDLE_MAY_THROW noexcept(false)
 #endif
@@ -180,76 +174,11 @@ struct BinaryCompareMessageConverter<false> {
 };
 }  // namespace details
 
-template <typename T>
-inline std::string ReplaceComplexTypeStr(std::string str,
-                                         const std::string& type_name) {
-  auto demangle_type_str = demangle(typeid(T).name());
-  size_t start_pos = 0;
-  while ((start_pos = str.find(demangle_type_str, start_pos)) !=
-         std::string::npos) {
-    str.replace(start_pos, demangle_type_str.length(), type_name);
-    start_pos += type_name.length();
-  }
-  return str;
-}
-
-#define __REPLACE_COMPLEX_TYPE_STR__(__TYPENAME, __STR)                      \
-  do {                                                                       \
-    __STR =                                                                  \
-        phi::enforce::ReplaceComplexTypeStr<__TYPENAME>(__STR, #__TYPENAME); \
-  } while (0)
-
-inline std::string SimplifyDemangleStr(std::string str) {
-  // the older is important, you have to put complex types in front
-  __REPLACE_COMPLEX_TYPE_STR__(paddle::framework::AttributeMap, str);
-  __REPLACE_COMPLEX_TYPE_STR__(paddle::framework::Attribute, str);
-  __REPLACE_COMPLEX_TYPE_STR__(paddle::imperative::NameVariableWrapperMap, str);
-  __REPLACE_COMPLEX_TYPE_STR__(paddle::imperative::NameVarBaseMap, str);
-  __REPLACE_COMPLEX_TYPE_STR__(std::string, str);
-  return str;
-}
-
-inline std::string GetCurrentTraceBackString(bool for_signal = false) {
-  std::ostringstream sout;
-
-  if (!for_signal) {
-    sout << "\n\n--------------------------------------\n";
-    sout << "C++ Traceback (most recent call last):";
-    sout << "\n--------------------------------------\n";
-  }
-#if !defined(_WIN32) && !defined(PADDLE_WITH_MUSL)
-  static constexpr int TRACE_STACK_LIMIT = 100;
-
-  void* call_stack[TRACE_STACK_LIMIT];
-  auto size = backtrace(call_stack, TRACE_STACK_LIMIT);
-  auto symbols = backtrace_symbols(call_stack, size);
-  Dl_info info;
-  int idx = 0;
-  // `for_signal` used to remove the stack trace introduced by
-  // obtaining the error stack trace when the signal error occurred,
-  // that is not related to the signal error self, remove it to
-  // avoid misleading users and developers
-  int end_idx = for_signal ? 2 : 0;
-  for (int i = size - 1; i >= end_idx; --i) {
-    if (dladdr(call_stack[i], &info) && info.dli_sname) {
-      auto demangled = demangle(info.dli_sname);
-      std::string path(info.dli_fname);
-      // C++ traceback info are from core.so
-      if (path.substr(path.length() - 3).compare(".so") == 0) {
-        sout << paddle::string::Sprintf(
-            "%-3d %s\n", idx++, SimplifyDemangleStr(demangled));
-      }
-    }
-  }
-  free(symbols);
-#else
-  sout << "Not support stack backtrace yet.\n";
-#endif
-  return sout.str();
-}
+std::string GetCurrentTraceBackString(bool for_signal = false);
+std::string SimplifyErrorTypeFormat(const std::string& str);
 
 template <typename StrType>
-inline std::string GetErrorSumaryString(StrType&& what,
+static std::string GetErrorSumaryString(StrType&& what,
                                         const char* file,
                                         int line) {
   std::ostringstream sout;
@@ -264,7 +193,20 @@ inline std::string GetErrorSumaryString(StrType&& what,
 }
 
 template <typename StrType>
-inline std::string GetTraceBackString(StrType&& what,
+std::string GetCompleteTraceBackString(StrType&& what,
+                                       const char* file,
+                                       int line) {
+  std::ostringstream sout;
+  sout << "\n----------------------\nError Message "
+          "Summary:\n----------------------\n";
+  sout << paddle::string::Sprintf(
+              "%s (at %s:%d)", std::forward<StrType>(what), file, line)
+       << std::endl;
+  return GetCurrentTraceBackString() + sout.str();
+}
+
+template <typename StrType>
+static std::string GetTraceBackString(StrType&& what,
                                       const char* file,
                                       int line) {
   if (FLAGS_call_stack_level > 1) {
@@ -273,19 +215,6 @@ inline std::string GetTraceBackString(StrType&& what,
   } else {
     return GetErrorSumaryString(what, file, line);
   }
-}
-
-inline std::string SimplifyErrorTypeFormat(const std::string& str) {
-  std::ostringstream sout;
-  size_t type_end_pos = str.find(":", 0);
-  if (type_end_pos == std::string::npos) {
-    sout << str;
-  } else {
-    // Remove "Error:", add "()""
-    sout << "(" << str.substr(0, type_end_pos - 5) << ")"
-         << str.substr(type_end_pos + 1);
-  }
-  return sout.str();
 }
 
 inline bool is_error(bool stat) { return !stat; }
@@ -347,6 +276,8 @@ struct EnforceNotMet : public std::exception {
       simple_err_str_ = str;
     }
   }
+
+  ~EnforceNotMet() override = default;
 
  private:
   // Used to determine the final type of exception thrown
