+++ paddle/phi/core/utils/type_registry.h
@@ -15,6 +15,7 @@ limitations under the License. */
 #pragma once
 
 #include <cassert>
+#include <map>
 #include <mutex>
 #include <string>
